hm
info depth 35 seldepth 100 score cp 145 time 1966802 nodes 70826295622 nps 36010880 pv d7d5 h2h3 h7h5 g4h5 e7e5 b1c3 b8c6 f1g2 g8f6 d2d3 c8e6 c1g5 d8d7 h5h6 f6h7 h6g7 f8g7 d1d2 f7f6 g5h6 g7h6 d2h6 e8c8 h1h2 d8g8 h3h4 d7d6 g1f3 h7g5 h4g5 h8h6 g5h6 e5e4 d3e4 d5e4
bestmove d7d5

Hits u32 max in iters before using all hash

+++ src/tree/half.rs
@@ -69,10 +69,10 @@ impl TreeHalf {
             end = start + block;
             self.next[thread].store(next + num, Ordering::Relaxed);
             self.end[thread].store(end, Ordering::Relaxed);
-            Some(NodePtr::new(self.half, start as u32))
+            Some(NodePtr::new(self.half, start))
         } else {
             self.next[thread].store(next + num, Ordering::Relaxed);
-            Some(NodePtr::new(self.half, next as u32))
+            Some(NodePtr::new(self.half, next))
         }
     }
